requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0